🚚 [variables] Move module `bindings` from `templates.domain`
