+++ src/cutty/entrypoints/cli/update.py
@@ -8,7 +8,7 @@ from cutty.entrypoints.cli.cookiecutter import extra_context_callback
 from cutty.entrypoints.cli.errors import fatal
 from cutty.projects.repository import ProjectRepository
 from cutty.services.update import update as service_update
-from cutty.templates.domain.bindings import Binding
+from cutty.variables.bindings import Binding
 
 
 @click.command()
