+++ src/cutty/projects/generate.py
@@ -6,7 +6,7 @@ from cutty.projects.project import Project
 from cutty.projects.template import Template
 from cutty.templates.adapters.questionary.prompts import createprompt
 from cutty.templates.domain.binders import bindvariables
-from cutty.templates.domain.bindings import Binding
+from cutty.variables.bindings import Binding
 
 
 def generate(
