+++ tests/unit/templates/adapters/cookiecutter/test_render.py
@@ -9,9 +9,9 @@ from cutty.filesystems.adapters.dict import DictFilesystem
 from cutty.filesystems.domain.path import Path
 from cutty.filesystems.domain.purepath import PurePath
 from cutty.templates.adapters.cookiecutter.render import createcookiecutterrenderer
-from cutty.templates.domain.bindings import Binding
 from cutty.templates.domain.config import Config
 from cutty.templates.domain.render import Renderer
+from cutty.variables.bindings import Binding
 
 
 @pytest.fixture
