+++ tests/unit/templates/adapters/questionary/test_prompts.py
@@ -10,7 +10,7 @@ from prompt_toolkit.output import DummyOutput
 
 from cutty.templates.adapters.questionary.prompts import createprompt
 from cutty.templates.domain.binders import Binder
-from cutty.templates.domain.bindings import GenericBinding
+from cutty.variables.bindings import GenericBinding
 from cutty.variables.variables import GenericVariable
 from tests.util.keys import BACKSPACE
 from tests.util.keys import ESCAPE
