+++ tests/unit/templates/domain/test_render.py
@@ -5,8 +5,8 @@ from cutty.filestorage.domain.files import Executable
 from cutty.filestorage.domain.files import RegularFile
 from cutty.filestorage.domain.files import SymbolicLink
 from cutty.filesystems.domain.purepath import PurePath
-from cutty.templates.domain.bindings import Binding
 from cutty.templates.domain.render import Renderer
+from cutty.variables.bindings import Binding
 from cutty.variables.variables import GenericVariable
