+++ tests/unit/templates/domain/test_renderfiles.py
@@ -4,9 +4,9 @@ import pytest
 from cutty.filestorage.domain.files import RegularFile
 from cutty.filesystems.adapters.dict import DictFilesystem
 from cutty.filesystems.domain.path import Path
-from cutty.templates.domain.bindings import Binding
 from cutty.templates.domain.render import Renderer
 from cutty.templates.domain.renderfiles import renderfiles
+from cutty.variables.bindings import Binding
 
 
 @pytest.fixture
